See #409 - still a work in progress. refactored the pixel handlers into seperate files and added test harnesses
