+++ tests/test_filereader.py
@@ -676,159 +676,6 @@ class ReadDataElementTests(unittest.TestCase):
         self.assertEqual(ds.DestinationAE, 'TEST  12')
 
 
-class JPEG_LS_Tests(unittest.TestCase):
-    def setUp(self):
-        self.jpeg_ls_lossless = read_file(jpeg_ls_lossless_name)
-        self.mr_small = read_file(mr_name)
-        self.emri_jpeg_ls_lossless = read_file(emri_jpeg_ls_lossless)
-        self.emri_small = read_file(emri_name)
-
-    def tearDown(self):
-        pydicom.config.force_gdcm_decompression = False
-        pydicom.config.force_pillow_decompression = False
-
-    def testGDCM_JPEG_LS_PixelArray(self):
-        """JPEG LS Lossless: Now works"""
-        pydicom.config.force_gdcm_decompression = True
-        pydicom.config.force_pillow_decompression = False
-        a = self.jpeg_ls_lossless.pixel_array
-        b = self.mr_small.pixel_array
-        self.assertEqual(a.mean(), b.mean(),
-                         "using GDCM Decoded pixel data is not all {0} (mean == {1})".format(b.mean(), a.mean()))
-
-    def testPillow_JPEG_LS_PixelArray(self):
-        pydicom.config.force_gdcm_decompression = False
-        pydicom.config.force_pillow_decompression = True
-        a = self.jpeg_ls_lossless.pixel_array
-        b = self.mr_small.pixel_array
-        self.assertEqual(a.mean(), b.mean(),
-                         "using Pillow Decoded pixel data is not all {0} (mean == {1})".format(b.mean(), a.mean()))
-
-    def test_emri_JPEG_LS_PixelArray(self):
-        """JPEG LS Lossless: Now works"""
-        if have_numpy and have_jpeg_ls:
-            a = self.emri_jpeg_ls_lossless.pixel_array
-            b = self.emri_small.pixel_array
-            self.assertEqual(a.mean(), b.mean(),
-                             "Decoded pixel data is not all {0} (mean == {1})".format(b.mean(), a.mean()))
-        else:
-            self.assertRaises(NotImplementedError, self.emri_jpeg_ls_lossless._get_pixel_array)
-
-
-class BigEndian_Tests(unittest.TestCase):
-    def setUp(self):
-        self.emri_big_endian = read_file(emri_big_endian_name)
-        self.emri_small = read_file(emri_name)
-
-    def test_big_endian_PixelArray(self):
-        """Test big endian pixel data vs little endian"""
-        if have_numpy:
-            a = self.emri_big_endian.pixel_array
-            b = self.emri_small.pixel_array
-            self.assertEqual(a.mean(), b.mean(),
-                             "Decoded big endian pixel data is not all {0} (mean == {1})".format(b.mean(), a.mean()))
-        else:
-            self.assertRaises(ImportError, self.emri_big_endian._get_pixel_array)
-
-
-class JPEG2000Tests(unittest.TestCase):
-    def setUp(self):
-        self.jpeg = read_file(jpeg2000_name)
-        self.jpegls = read_file(jpeg2000_lossless_name)
-        self.mr_small = read_file(mr_name)
-        self.emri_jpeg_2k_lossless = read_file(emri_jpeg_2k_lossless)
-        self.emri_small = read_file(emri_name)
-
-    def testJPEG2000(self):
-        """JPEG2000: Returns correct values for sample data elements............"""
-        expected = [Tag(0x0054, 0x0010), Tag(0x0054, 0x0020)]  # XX also tests multiple-valued AT data element
-        got = self.jpeg.FrameIncrementPointer
-        self.assertEqual(got, expected, "JPEG2000 file, Frame Increment Pointer: expected %s, got %s" % (expected, got))
-
-        got = self.jpeg.DerivationCodeSequence[0].CodeMeaning
-        expected = 'Lossy Compression'
-        self.assertEqual(got, expected, "JPEG200 file, Code Meaning got %s, expected %s" % (got, expected))
-
-    def testJPEG2000PixelArray(self):
-        """JPEG2000: Now works"""
-        if have_numpy and have_pillow:
-            a = self.jpegls.pixel_array
-            b = self.mr_small.pixel_array
-            self.assertEqual(a.mean(), b.mean(),
-                             "Decoded pixel data is not all {0} (mean == {1})".format(b.mean(), a.mean()))
-        else:
-            self.assertRaises(NotImplementedError, self.jpegls._get_pixel_array)
-
-    def test_emri_JPEG2000PixelArray(self):
-        """JPEG2000: Now works"""
-        if have_numpy and have_pillow:
-            a = self.emri_jpeg_2k_lossless.pixel_array
-            b = self.emri_small.pixel_array
-            self.assertEqual(a.mean(), b.mean(),
-                             "Decoded pixel data is not all {0} (mean == {1})".format(b.mean(), a.mean()))
-        else:
-            self.assertRaises(NotImplementedError, self.emri_jpeg_2k_lossless._get_pixel_array)
-
-
-class JPEGlossyTests(unittest.TestCase):
-
-    def setUp(self):
-        self.jpeg = read_file(jpeg_lossy_name)
-        self.color_3d_jpeg = read_file(color_3d_jpeg_baseline)
-
-    def testJPEGlossy(self):
-        """JPEG-lossy: Returns correct values for sample data elements.........."""
-        got = self.jpeg.DerivationCodeSequence[0].CodeMeaning
-        expected = 'Lossy Compression'
-        self.assertEqual(got, expected, "JPEG-lossy file, Code Meaning got %s, expected %s" % (got, expected))
-
-    def testJPEGlossyPixelArray(self):
-        """JPEG-lossy: Fails gracefully when uncompressed data is asked for....."""
-        if have_pillow and have_numpy:
-            self.assertRaises(NotImplementedError, self.jpeg._get_pixel_array)
-        else:
-            self.assertRaises(NotImplementedError, self.jpeg._get_pixel_array)
-
-    def testJPEGBaselineColor3DPixelArray(self):
-        if have_pillow and have_numpy:
-            a = self.color_3d_jpeg.pixel_array
-            self.assertEqual(a.shape, (120, 480, 640, 3))
-            # this test points were manually identified in Osirix viewer
-            self.assertEqual(tuple(a[3, 159, 290, :]), (41, 41, 41))
-            self.assertEqual(tuple(a[3, 169, 290, :]), (57, 57, 57))
-        else:
-            self.assertRaises(NotImplementedError, self.color_3d_jpeg._get_pixel_array)
-
-
-class JPEGlosslessTests(unittest.TestCase):
-    def setUp(self):
-        self.jpeg = read_file(jpeg_lossless_name)
-
-    def testJPEGlossless(self):
-        """JPEGlossless: Returns correct values for sample data elements........"""
-        got = self.jpeg.SourceImageSequence[0].PurposeOfReferenceCodeSequence[0].CodeMeaning
-        expected = 'Uncompressed predecessor'
-        self.assertEqual(got, expected, "JPEG-lossless file, Code Meaning got %s, expected %s" % (got, expected))
-
-    def testJPEGlosslessPixelArray(self):
-        """JPEGlossless: Fails gracefully when uncompressed data is asked for..."""
-        # This test passes if the call raises either an
-        # ImportError when there is no Pillow module
-        # Or
-        # NotImplementedError when there is a Pillow module
-        #    but it lacks JPEG Lossless Dll's
-        # Or
-        # the call does not raise any Exceptions
-        # This test fails if any other exception is raised
-        with self.assertRaises((ImportError, NotImplementedError)):
-            try:
-                _x = self.jpeg._get_pixel_array()
-            except Exception:
-                raise
-            else:
-                raise ImportError()
-
-
 class DeferredReadTests(unittest.TestCase):
     """Test that deferred data element reading (for large size)
     works as expected
